requirements.txt
opencv_python
Flask==1.0.2
Pillow==6.0.0
torchvision==0.2.1
numpy==1.15.4
gunicorn==19.9.0

